+++ src/vue/modules/assets/my-assets-explorer/index.vue
@@ -49,7 +49,7 @@
 
       <div class="my-assets-explorer__asset-list-wrp">
         <div
-          v-if="assets.length"
+          v-if="accountOwnedAssetsBalances.length"
           class="my-assets-explorer__asset-list"
         >
           <template v-for="item in accountOwnedAssetsBalances">
@@ -138,7 +138,6 @@ export default {
 
   computed: {
     ...mapGetters({
-      assets: vuexTypes.balancesAssets,
       ownedAssets: vuexTypes.ownedBalancesAssets,
       accountBalances: vuexTypes.accountBalances,
       accountOwnedAssetsBalances: vuexTypes.accountOwnedAssetsBalances,