+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Fixed
+- Customer UI switch is shown only for corporate accounts now
+- Change Customer UI => Customer in EN locale
+
 ## [1.10.0-rconto.3] - 2019-07-23
 #### Added
 - Simultaneous support for most common delimiters (spaces, tabs, colons, commas,