+++ CHANGELOG.md
@@ -11,6 +11,7 @@ for further information about branching and tagging conventions.
 #### Fixed
 - A bug when you cannot return from "Downloads" if the tab was opened in new tab
 - A bug when page reload twice
+- A bug when user cannot sign out
 
 ### "Under the hood" changes
 #### Added