+++ CHANGELOG.md
@@ -12,6 +12,9 @@ for further information about branching and tagging conventions.
 - Customer UI switch is shown only for corporate accounts now
 - Change Customer UI => Customer in EN locale
 
+#### Changed
+- Required only 'Business name' field in business kyc form
+
 ## [1.10.0-rconto.3] - 2019-07-23
 #### Added
 - Simultaneous support for most common delimiters (spaces, tabs, colons, commas,