+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Changed
+- Display asset description completely
+- Required only 'Business name' field in business kyc form
+
+#### Fixed
+- Resolved an issue with randomly shown/hidden sidebar menu links of corporate
+  account
+
 ## [1.0.0-rc.0] - 2019-07-25
 ### Added
 - New 'Explore atomic swaps' page